CFIN-215 Refactor methods to format responses
